Clean: Rearrange the ql_packager modules
