+++ crates/ql_packager/src/lib.rs
@@ -10,6 +10,7 @@ use thiserror::Error;
 
 use ql_instances::DownloadError;
 
+mod export;
 mod import;
 mod multimc;
